registratration added with  password hashing , TODO hashing when edit user
commented out hash query's in edit_user , come back later

+++ admin/includes/view_all_posts.php
@@ -65,13 +65,24 @@
       echo "<a href='posts.php?source=edit_post&p_id={$post_id}' class='';>Edit</a></td>";
 
       echo "<td style='display: flex;justify-content: center;align-items: center;background: red;'>";
-      echo "<a href='posts.php?delete={$post_id}' class='';>Delete</a></td>";
+      echo "<a href='posts.php?delete={$post_id}' class=''; onclick=\" return confirm('Are you sure you want to delete?'); \">Delete</a></td>";
       
       echo "</tr>";
     }
 
 
     ?>
+<script>
+function confirmMe() {
+//TODO write better confirm function!!
+
+confirm('are you sure you want to delete?');
+
+return
+
+}
+</script>
+
 
   </tbody>
